Factor out nominal flow data reader
